Fix APICreatePod and fix up some programmatic errors in REST route
