+++ rest/route/pod_agent_test.go
@@ -1,7 +1,6 @@
 package route
 
 import (
-	"bytes"
 	"context"
 	"net/http"
 	"testing"
@@ -9,10 +8,8 @@ import (
 	"github.com/evergreen-ci/evergreen"
 	"github.com/evergreen-ci/evergreen/apimodels"
 	"github.com/evergreen-ci/evergreen/mock"
-	"github.com/evergreen-ci/evergreen/model/pod"
 	"github.com/evergreen-ci/evergreen/rest/data"
 	"github.com/evergreen-ci/gimlet"
-	"github.com/evergreen-ci/utility"
 	"github.com/mongodb/grip/send"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -203,157 +200,6 @@ func TestPodAgentNextTask(t *testing.T) {
 			require.True(t, ok)
 
 			tCase(ctx, t, rh, env)
-		})
-	}
-}
-
-func TestPostPod(t *testing.T) {
-	for tName, tCase := range map[string]func(ctx context.Context, t *testing.T, ph *podPostHandler){
-		"FactorySucceeds": func(ctx context.Context, t *testing.T, ph *podPostHandler) {
-			copied := ph.Factory()
-			assert.NotZero(t, copied)
-			_, ok := copied.(*podPostHandler)
-			assert.True(t, ok)
-		},
-		"ParseSucceeds": func(ctx context.Context, t *testing.T, ph *podPostHandler) {
-			json := []byte(`{
-				"memory": 128,
-				"cpu": 128,
-				"image": "image",
-				"env_vars": [{
-					"name": "env_name",
-					"value": "env_val",
-					"secret": false
-				}],
-				"platform": "linux",
-				"secret": "secret"
-			}`)
-			req, err := http.NewRequest(http.MethodPut, "https://example.com/rest/v2/pods/123abc", bytes.NewBuffer(json))
-			require.NoError(t, err)
-			require.NoError(t, ph.Parse(ctx, req))
-			assert.Equal(t, 128, utility.FromIntPtr(ph.p.Memory))
-			assert.Equal(t, 128, utility.FromIntPtr(ph.p.CPU))
-			assert.Equal(t, "image", utility.FromStringPtr(ph.p.Image))
-			assert.Equal(t, "linux", utility.FromStringPtr(ph.p.Platform))
-			assert.Equal(t, "secret", utility.FromStringPtr(ph.p.Secret))
-			assert.Equal(t, "env_name", utility.FromStringPtr(ph.p.EnvVars[0].Name))
-			assert.Equal(t, "env_val", utility.FromStringPtr(ph.p.EnvVars[0].Value))
-			assert.Equal(t, false, utility.FromBoolPtr(ph.p.EnvVars[0].Secret))
-		},
-		"ParseSucceedsWithSecrets": func(ctx context.Context, t *testing.T, ph *podPostHandler) {
-			json := []byte(`{
-				"memory": 128,
-				"cpu": 128,
-				"image": "image",
-				"env_vars": [{
-					"name": "secret_name",
-					"value": "secret_val",
-					"secret": true
-				}],
-				"platform": "linux",
-				"secret": "secret"
-			}`)
-			req, err := http.NewRequest(http.MethodPut, "https://example.com/rest/v2/pods/123abc", bytes.NewBuffer(json))
-			require.NoError(t, err)
-			require.NoError(t, ph.Parse(ctx, req))
-			assert.Equal(t, "secret_name", utility.FromStringPtr(ph.p.EnvVars[0].Name))
-			assert.Equal(t, "secret_val", utility.FromStringPtr(ph.p.EnvVars[0].Value))
-			assert.Equal(t, true, utility.FromBoolPtr(ph.p.EnvVars[0].Secret))
-		},
-		"ParseFailsWithInvalidJSON": func(ctx context.Context, t *testing.T, ph *podPostHandler) {
-			json := []byte(`{
-				"memory": 128,
-				"cpu": 128,,,
-				"image": "image",
-				"env_vars": [],
-				"platform": "linux",
-				"secret": "secret"
-			}`)
-			req, err := http.NewRequest(http.MethodPut, "https://example.com/rest/v2/pods/123abc", bytes.NewBuffer(json))
-			require.NoError(t, err)
-			require.Error(t, ph.Parse(ctx, req))
-		},
-		"ParseFailsWithMissingInput": func(ctx context.Context, t *testing.T, ph *podPostHandler) {
-			json := []byte(`{
-				"cpu": 128,
-				"image": "image",
-				"env_vars": [],
-				"platform": "linux",
-				"secret": "secret"
-			}`)
-			req, err := http.NewRequest(http.MethodPut, "https://example.com/rest/v2/pods/123abc", bytes.NewBuffer(json))
-			require.NoError(t, err)
-			require.Error(t, ph.Parse(ctx, req))
-		},
-		"ParseFailsWithMissingEnvVarInput": func(ctx context.Context, t *testing.T, ph *podPostHandler) {
-			json := []byte(`{
-				"cpu": 128,
-				"image": "image",
-				"env_vars": [{
-					"value": "secret_val",
-					"secret": true
-				}],
-				"platform": "linux",
-				"secret": "secret"
-			}`)
-			req, err := http.NewRequest(http.MethodPut, "https://example.com/rest/v2/pods/123abc", bytes.NewBuffer(json))
-			require.NoError(t, err)
-			require.Error(t, ph.Parse(ctx, req))
-		},
-		"ParseFailsWithInvalidInput": func(ctx context.Context, t *testing.T, ph *podPostHandler) {
-			json := []byte(`{
-				"memory": -1,
-				"cpu": 128,
-				"image": "image",
-				"env_vars": [],
-				"platform": "linux",
-				"secret": "secret"
-			}`)
-			req, err := http.NewRequest(http.MethodPut, "https://example.com/rest/v2/pods/123abc", bytes.NewBuffer(json))
-			require.NoError(t, err)
-			require.Error(t, ph.Parse(ctx, req))
-		},
-		"RunSucceedsWithValidInput": func(ctx context.Context, t *testing.T, ph *podPostHandler) {
-			json := []byte(`{
-				"memory": 128,
-				"cpu": 128,
-				"image": "image",
-				"env_vars": [{
-					"name": "env_name",
-					"value": "env_val",
-					"secret": false
-				}],
-				"platform": "linux",
-				"secret": "secret"
-			}`)
-			ph.body = json
-
-			resp := ph.Run(ctx)
-			require.NotNil(t, resp.Data())
-			assert.Equal(t, http.StatusCreated, resp.Status())
-		},
-	} {
-		t.Run(tName, func(t *testing.T) {
-			ctx, cancel := context.WithCancel(context.Background())
-			defer cancel()
-
-			sc := &data.MockConnector{
-				MockPodConnector: data.MockPodConnector{
-					CachedPods: []pod.Pod{
-						{
-							ID: "pod1",
-						},
-						{
-							ID: "pod2",
-						},
-					},
-				},
-			}
-
-			p := makePostPod(sc)
-			require.NotZero(t, p)
-
-			tCase(ctx, t, p.(*podPostHandler))
 		})
 	}
 }
